+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+### Changed
+
+- Add Helm annotations to namespace resource to allow namespaces to be managed
+in charts.
+
 ## [2.2.1] - 2020-08-19
 
 ### Changed